Version two
This version of the project is based on an idea I had to make this survey actually look like a phone. This is not a responsive design, includes more elements than the first version, and needs some refactoring in terms of margins, padding, and alignment. It looks the way I wanted it to look, but does not do so cleanly at all times.